pyproject.toml
[project]
name = "django-cafeteria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-tailwind>=1.0.3",
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "ipython>=9.3.0",
    "pillow>=11.2.1",
]
dev-dependencies = [
    "ipython>=9.3.0",
]
